# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk aplikasi Anda",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/manager": {
            "get": {
                "tags": [
                    "Managers"
                ],
                "summary": "Menampilkan daftar semua manager",
                "description": "Endpoint ini digunakan untuk mendapatkan daftar manager berdasarkan perusahaan yang dimiliki oleh pengguna saat ini.",
                "operationId": "e5b35d2942aa68747cec3363d92559ed",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Nama manager untuk pencarian",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Jane Doe"
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Field yang digunakan untuk pengurutan",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "name"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Urutan pengurutan (asc/desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar manager",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/{id}": {
            "get": {
                "tags": [
                    "Managers"
                ],
                "summary": "Menampilkan detail manager",
                "description": "Endpoint ini digunakan untuk mendapatkan detail manager berdasarkan ID.",
                "operationId": "402d360b2872e467824229e4086c018d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID manager",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan detail manager",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Manager tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Manager not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Managers"
                ],
                "summary": "Menampilkan profil pengguna saat ini",
                "description": "Endpoint ini digunakan untuk mendapatkan data profil pengguna yang sedang login.",
                "operationId": "b2f25d7d05f551dcf652724d75f57186",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan profil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Managers"
                ],
                "summary": "Memperbarui profil pengguna",
                "description": "Endpoint ini digunakan untuk memperbarui profil pengguna yang sedang login.",
                "operationId": "d8bc72fea68cb42eaf036d129eac7bf4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "address"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Contoh Baru No. 123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}